parseq/requirements/tune.txt
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
attrs==23.1.0
ax-platform==0.3.3
backcall==0.2.0
botorch==0.8.5
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.4
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.7
executing==1.2.0
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
gpytorch==1.10
grpcio==1.43.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-resources==6.0.1
ipython==8.12.2
ipywidgets==8.1.0
jedi==0.19.0
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lazy-loader==0.3
linear-operator==0.4.0
lmdb==1.4.1
markupsafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multipledispatch==1.0.0
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.16.1
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.6
python-dateutil==2.8.2
pytz==2023.3
pywavelets==1.4.1
pyyaml==6.0.1
ray==1.13.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tensorboardx==2.6.2
threadpoolctl==3.2.0
tifffile==2023.7.10
tqdm==4.66.1
traitlets==5.9.0
typeguard==2.13.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.3
wcwidth==0.2.6
widgetsnbextension==4.0.8
zipp==3.16.2
